+++ src/test/kotlin/cl/ravenhill/jakt/constraints/pairs/BeStrictlyOrderedTest.kt
@@ -1,9 +1,0 @@
-package cl.ravenhill.jakt.constraints.pairs
-
-import io.kotest.core.spec.style.FreeSpec
-import io.kotest.matchers.shouldBe
-
-class BeStrictlyOrderedTest : FreeSpec({
-
-
-})